+++ src/ops_data_store/cli/db/__init__.py
@@ -1,3 +1,4 @@
+import logging
 from pathlib import Path
 from typing import Annotated
 
@@ -9,6 +10,26 @@ from ops_data_store.config import Config
 app = typer.Typer()
 
 config = Config()
+logger = logging.getLogger("app")
+
+
+@app.command(help="Check DB connectivity.")
+def check() -> None:
+    """Check DB connection."""
+    try:
+        logger.info("Checking DB connectivity.")
+        with psycopg.connect(conninfo=config.DB_DSN, connect_timeout=3) as conn:
+            conn.read_only = True
+
+            with conn.cursor() as cur:
+                cur.execute("SELECT 1")
+
+            logger.info("DB connectivity ok.")
+            print("Ok. DB connection successful.")
+    except (psycopg.ProgrammingError, psycopg.OperationalError) as e:
+        logger.error(e, exc_info=True)
+        print("No. DB connection failed.")
+        raise typer.Abort() from e
 
 
 @app.command(help="Execute contents of an SQL against DB.")